Remove v2.0.0, v2.5.0, Rename v2.5.1, v2.6.0, and add v3.0 adapter

+++ ccore/nebula/types/driver.go
@@ -46,7 +46,7 @@ type (
 
 	ExecutionResponse interface {
 		GetErrorCode() nerrors.ErrorCode
-		GetLatencyInUs() int32
+		GetLatencyInUs() int64
 		GetData() DataSet
 		GetSpaceName() []byte
 		GetErrorMsg() []byte
